+++ lib/smnoisebandpartition.hh
@@ -5,6 +5,8 @@
 
 #include <vector>
 
+#include <stdint.h>
+
 #include "smrandom.hh"
 
 namespace SpectMorph
@@ -18,7 +20,7 @@ class NoiseBandPartition
 
 public:
   NoiseBandPartition (size_t n_bands, size_t n_spectrum_bins, double mix_freq);
-  void noise_envelope_to_spectrum (SpectMorph::Random& random_gen, const std::vector<float>& envelope, float *spectrum, double scale);
+  void noise_envelope_to_spectrum (SpectMorph::Random& random_gen, const std::vector<uint16_t>& envelope, float *spectrum, double scale);
 
   size_t n_bands();
   size_t n_spectrum_bins();
